fix:remove redundant syntax
